move dashboard files to a new module
