+++ dashboard/main.py
@@ -8,12 +8,11 @@ import pandas as pd
 import streamlit as st
 from fortepyan import MidiPiece
 from omegaconf import OmegaConf
-from datasets import load_dataset
+from predict_piece import predict_piece_dashboard
 
 from model import make_model
 from data.quantizer import MidiQuantizer
 from data.dataset import load_cache_dataset
-from predict_piece import predict_piece_dashboard
 from utils import vocab_sizes, piece_av_files, generate_sequence
 
 # For now let's run all dashboards on CPU
@@ -22,10 +21,8 @@ DEVICE = "cpu"
 
 def main():
     with st.sidebar:
-        mode = st.selectbox(label="Display", options=["Model predictions", "Predict piece", "Tokenization review"])
+        mode = st.selectbox(label="Display", options=["Model predictions", "Predict piece"])
 
-    if mode == "Tokenization review":
-        tokenization_review_dashboard()
     if mode == "Predict piece":
         predict_piece_dashboard()
     if mode == "Model predictions":
@@ -186,57 +183,6 @@ def model_predictions_review():
             # Predicted
             st.image(predicted_paths["pianoroll_path"])
             st.audio(predicted_paths["mp3_path"])
-
-
-def tokenization_review_dashboard():
-    st.markdown("### Quantization settings")
-
-    n_dstart_bins = st.number_input(label="n_dstart_bins", value=3)
-    n_duration_bins = st.number_input(label="n_duration_bins", value=3)
-    n_velocity_bins = st.number_input(label="n_velocity_bins", value=3)
-
-    quantizer = MidiQuantizer(
-        n_dstart_bins=n_dstart_bins,
-        n_duration_bins=n_duration_bins,
-        n_velocity_bins=n_velocity_bins,
-    )
-
-    split = "train"
-    dataset_name = "roszcz/maestro-v1-sustain"
-    dataset = load_dataset(dataset_name, split=split)
-
-    cols = st.columns(2)
-    with cols[0]:
-        st.markdown("### Unchanged")
-    with cols[1]:
-        st.markdown("### Quantized")
-
-    np.random.seed(137)
-    n_samples = 5
-    idxs = np.random.randint(len(dataset), size=n_samples)
-    for idx in idxs:
-        piece = MidiPiece.from_huggingface(dataset[int(idx)])
-        quantized_piece = quantizer.quantize_piece(piece)
-
-        av_dir = "tmp/dashboard/common"
-        bins = f"{n_dstart_bins}-{n_duration_bins}-{n_velocity_bins}"
-        save_name = f"{dataset_name}-{split}-{idx}".replace("/", "_")
-
-        save_base_gt = os.path.join(av_dir, f"{save_name}")
-        gt_paths = piece_av_files(piece, save_base=save_base_gt)
-
-        save_base_quantized = os.path.join(av_dir, f"{save_name}-{bins}")
-        quantized_piece_paths = piece_av_files(quantized_piece, save_base=save_base_quantized)
-
-        st.json(piece.source)
-        cols = st.columns(2)
-        with cols[0]:
-            st.image(gt_paths["pianoroll_path"])
-            st.audio(gt_paths["mp3_path"])
-
-        with cols[1]:
-            st.image(quantized_piece_paths["pianoroll_path"])
-            st.audio(quantized_piece_paths["mp3_path"])
 
 
 def prepare_midi_pieces(
